requirements-colab-ct.txt
hydra-core==1.3.2
omegaconf==2.3.0
piq==0.8.0
prettytable==3.14.0
setproctitle==1.3.3
wandb==0.17.8
imageio==2.37.0
diffusers>=0.30.0
accelerate>=0.30.0
pydicom==3.0.1
safetensors==0.5.2
scipy==1.14.1
